backend/requirements.txt
pytest
httpx
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
pytest-asyncio
azure-identity
azure-core
azure-keyvault-secrets
validators 
orjson
pylint
mypy
dotenv
scrapy
sqlmodel
psycopg2-binary
